Complete employee applying SRP
